Move MonoBehaviours into their own folder and namespace
