+++ Seeds of Love/Assets/Script/Behaviour.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: afb06ca39e9119647a196b8707d4d1b2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
