+++ Seeds of Love/Assets/Script/Behaviour/Detector.cs.meta
